Added start screen and completed more animations

+++ GXPEngine/GXPEngine/Sprite.cs
@@ -135,7 +135,7 @@ namespace GXPEngine
 		/// <summary>
 		/// Gets or sets the sprite's width in pixels.
 		/// </summary>
-		virtual public int width {
+		virtual public float width {
 			get { 
 				if (_texture != null) return (int)Math.Abs(_texture.width * _scaleX);
 				return 0;
@@ -151,7 +151,7 @@ namespace GXPEngine
 		/// <summary>
 		/// Gets or sets the sprite's height in pixels.
 		/// </summary>
-		virtual public int height {
+		virtual public float height {
 			get { 
 				if (_texture != null) return (int)Math.Abs(_texture.height * _scaleY);
 				return 0;
